fix: #14 Facade Patter Apply Fonts Get And Compared-Fonts Preview
