+++ src/main/java/sopt/noonnu/facade/FontFacadeService.java
@@ -4,11 +4,18 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import sopt.noonnu.font.domain.Font;
+import sopt.noonnu.font.dto.command.GetFontsCommand;
+import sopt.noonnu.font.dto.response.FontListResponse;
+import sopt.noonnu.font.dto.response.FontPreviewListResponse;
 import sopt.noonnu.font.service.FontService;
 import sopt.noonnu.user.domain.User;
 import sopt.noonnu.user.service.UserService;
+import sopt.noonnu.userfont.domain.UserFonts;
 import sopt.noonnu.userfont.dto.command.UpdateFontFlagCommandDto;
 import sopt.noonnu.userfont.service.UserFontService;
+
+import java.util.List;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -30,5 +37,33 @@ public class FontFacadeService {
         Font font = fontService.getFont(command.fontId());
         User user = userService.getUser(command.userId());
         userFontService.updateCompareFont(font, user, command.isTrue());
+    }
+
+    @Transactional(readOnly = true)
+    public FontListResponse getFonts(GetFontsCommand command){
+        List<Font> fonts = fontService.getFonts(command);
+
+        Map<Long, UserFonts> userFontMap = userFontService.getUserFontMapByUserId(command.userId());
+
+        List<FontListResponse.FontResponse> fontResponses = fonts.stream()
+                .map(font -> {
+                    Long fontId = font.getId();
+                    UserFonts userFont = userFontMap.get(fontId);
+
+                    boolean isLiked = userFont != null && userFont.isLiked();
+                    boolean isCompared = userFont != null && userFont.isCompared();
+
+                    return FontListResponse.FontResponse.of(font, isLiked, isCompared);
+                })
+                .toList();
+
+        return FontListResponse.from(fontResponses);
+    }
+
+    @Transactional(readOnly = true)
+    public FontPreviewListResponse getComparedFontPreviews(Long userId) {
+        List<Font> fonts = userFontService.getComparedFontPreviews(userId);
+
+        return FontPreviewListResponse.from(fonts);
     }
 }
